+++ Src/SOS.Observations.Api/Controllers/AreasController.cs
@@ -55,22 +55,22 @@ namespace SOS.Observations.Api.Controllers
 
 
         /// <inheritdoc />
-        [HttpGet("{areaType}/{feature}/Export")]
+        [HttpGet("{areaType}/{featureId}/Export")]
         [ProducesResponseType(typeof(byte[]), (int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.InternalServerError)]
         [ProducesResponseType((int)HttpStatusCode.NoContent)]
-        public async Task<IActionResult> ExportAreaAsync([FromRoute] AreaTypeDto areaType, [FromRoute] string feature)
+        public async Task<IActionResult> ExportAreaAsync([FromRoute] AreaTypeDto areaType, [FromRoute] string featureId)
         {
             try
             {
-                var zipBytes = await _areaManager.GetZipppedAreaAsync(areaType, feature);
+                var zipBytes = await _areaManager.GetZipppedAreaAsync(areaType, featureId);
 
                 if (zipBytes == null)
                 {
                     return new StatusCodeResult((int)HttpStatusCode.NoContent);
                 }
 
-                return File(zipBytes, "application/zip", $"Area{areaType:G}:{feature}.zip");
+                return File(zipBytes, "application/zip", $"Area{areaType:G}:{featureId}.zip");
             }
             catch (Exception e)
             {